update/pandas_export.py
made the test loop into a function for better ease of editing between a team, moved onto exporting all integers by iterations in ascending order including the raw data in a pandas dataframe to observe any potential iteration cluster patterns. then i began to parse the final number of each integer in that list (in the same order) to see if there is a pattern there. this will require more analysis and better visualization tools.